sky: rework sky shader
The sky shader now interpolates the sky gradient differently, which
should better match the original.
The sun is faded-out more smoothly based on radius.
The background texture is affected by background light. In reality it
should be affected by fog color, but that will be done later.
Added sun radius and additive effect.